backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
apscheduler==3.10.4
python-dateutil==2.8.2
pytest==7.4.4
pytest-asyncio==0.23.3
pandas==2.1.4
